Feature/864e6n8bj change string for numeric datatype (#3)
* Change datatype for financial data entity String for double

* Add try-catch for number format exception for csv reader

* Change data type for prices in consumer service

* remove unnecessary price setter

* Add liquibase log

* Change liquibase datatype

+++ services/LocalStorageMarketData/src/main/java/market/localstoragemarketdata/data/generator/RandomWalkGeneratorScheduler.java
@@ -8,6 +8,7 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.ObjectOutputStream;
 import java.lang.instrument.Instrumentation;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Random;
 
@@ -36,7 +37,7 @@ public class RandomWalkGeneratorScheduler {
                 randomTimeDistance = random.nextInt(MIN_TIME_DISTANCE_BETWEEN_TICKS, MAX_TIME_DISTANCE_BETWEEN_TICKS);
 
 
-                float value = randomWalkGenerator.generateRandomValue();
+                double value = randomWalkGenerator.generateRandomValue();
                 LocalDateTime date = randomWalkGenerator.generateRandomLocalDatetime();
                 randomWalkGenerator.loadDataToInMemoryStorage(date, value);
